# Speed up bootstrap with large number of tokens in the cluster 10X