@scylladb-promoter 2022-06-22
CI state **`SUCCESS`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/989/