@slivne 2022-07-14
Is this only for testing
- simple_strategy is non production - please check with network strategy 
@kbr- please review as well