@asias 2022-07-15
> Is this only for testing
> 
>     * simple_strategy is non production - please check with network strategy
>       @kbr- please review as well

This is a fix for two users in production. They are waiting for the fix. We have tables use simple_strategy by default. This issue preventing them to replace a dead node.