@scylladb-promoter 2022-07-19
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/1416/