@asias 2022-07-26
Changes in 2924a9e3b76b3a1c6603d3a249dcf52ec399a45f:
- move the is every strategy check outside the loop