@asias 2022-07-26
Changes in 6152f5b858d0cbd9545ac5558128fd283ddf3d4b
- Switch to use unordered_set in update_normal_token_owners